Change #9.1
- Very minor changes this time, not sure what was causing variable
issues on 5.12~5.1211, couldn’t find the source of the issue, manually
set the variable value on that set of pages. Seems to be functioning
this way.
- The textScreen[]s were all one higher than they should have been from
screenText[29] onwards, so lowered all by 1 to fix the wrong text
showing up
- Experiment with setting variables rather than having the screen
increase it by one—does this work? And, if it does, why did I note use
this initially?