Change #9.2
- My world has been rocked. The variables are needed in order to stop 2
buttons being pressed at once, but the variable += 1 code? Entirely.
Unneeded. All of goWest += 1 (besides the necessary one, which is the
multiple clicks of the same buttons on Screen 2.4, I believe) has been
removed, and the same is being done for the other variables.

I need a glass of water. My mind has been blown.