Change #12
- Expanded on drug trip path and added a screen to Go East
- Replaced textBoxX & textBoxY with textBoxCoordinates, as both
variables were = 30
- Put (and commented out) DrawButtons under ScreenSetup, too cumbersome
to have it working at the moment. Will be uncommented after all screens
use DrawButtons and don’t manually draw

Current Yearnings:

- Implement an array, for crying out loud! Though the zombie path was
scrapped for the time being, a satirical path solely for utilizing
arrays might not be a bad idea?
- Do I even have to say “more work”?
- Make sure to stay away from shrooms, the drug trip path illustrated
the dangers of it
- Storyboard?
- Prioritize loading strings from a text file and/or getting
ScreenSetup to include String screenText[] (is this possible?)
- Brainstorm ways to cut down on code, and more importantly weed out
unnecessary code (the first of many, for instance, was cutting the
resetting of variables out of resetting to screen 1.0—not only does
this if(screen == 1.0) anyways, but it is also set per screen at this
point.
- Edit the writing, be a funnier person all around